# Speed up process_crosslinks(...) and get_crosslink_deltas(...) by 10x - 15x in state_sim