Successor PR to https://github.com/status-im/nim-beacon-chain/pull/313 (issues with rebasing conflicts).

This switches inner/outer loop nesting order to get 10-15x function speedup for 128 and 512 validator cases by avoiding accidentally quadratic behavior, while keeping function signature unchanged and allowing easy ongoing verification of correctness of optimization.

This takes `process_crosslinks(...)` processing times down in my testing from 15-16s or so to 1s and from 22-24s to 1.5s. `get_crosslink_deltas(...)` sees similar speedups, since it was also bottlenecked on `get_winning_crosslink_and_attesting_indices(...)`.

The other big epoch time-consumer is `get_attestation_deltas(...)`, which https://github.com/status-im/nim-beacon-chain/pull/314/commits/4590b697abd71fce26f17d39eeffe6a64994e4d0 decreases by 90x or so for me for 512 validators, from 178s to 2s.

Timings I get with 512 validators:
```
All time are ms
     Average,       StdDev,          Min,          Max,      Samples,         Test
     197.915,       46.600,      118.005,      309.630,          128, Process non-epoch slot with block
    3224.563,     2258.988,     1627.217,     4821.908,            2, Process epoch slot with block
       1.997,        1.181,        0.027,        4.247,          130, Tree-hash block
       8.371,        0.291,        7.516,        9.636,          130, Retrieve committee once using get_crosslink_committee
      64.856,       18.607,       31.645,      113.677,         8320, Combine committee attestations
```