@tersec 2019-07-08
> So I suppose the main gain was due to lots of seq allocations in inner loops?
> 
> There are a couple of GC profiling information available according to https://github.com/nim-lang/Nim/blob/devel/lib/system/gc.nim:
> 
>     * `-d:memProfiler`
> 
>     * `GC_getStatistics()`
> 
> 
> It would be interesting to understand how we were stressing th GC before and after this change.

There are some distinct cases:

https://github.com/status-im/nim-beacon-chain/pull/314/commits/d8f63d256653d20acf236f5111ab645ca4ef06b3 I don't think is mostly working by memory allocations per se (though, it does help). Rather, there were lots of `filterIt` scans with very sparse results in an inner loop.

https://github.com/status-im/nim-beacon-chain/pull/314/commits/61f388f1d9354a5ecc0a74ce9fadb0298cd58e19 is similar:
```
        get_attesting_balance(
          state,
          filterIt(attestations, it.data.crosslink == candidate_crosslink),
          stateCache)
```

wasn't matching most attestations most times, but it was scanning them, and it was O(n^2) overall, because number of attestations is proportional to number of crosslinks. So this turns that into O(n) by avoiding repetitious scans.

The `seq` to `openarray` change in that commit is nice, but perf-wise negligible.

https://github.com/status-im/nim-beacon-chain/pull/314/commits/776833ec9d7c0e436821f5b41d17373a8d26fdc5 actually adds memory allocation/churn. I put it there because the spec does and if it turns out to be problematic or unnecessary, it's an obvious target for removal. In particular, I don't see how it could ever be doing much useful -- it's basically deduplicating a list, but that list is internally generated by us, in a way that if there are duplicates, is already a bug. Because it's a set, order doesn't matter for the spec, so that shouldn't be a problem.

I think https://github.com/status-im/nim-beacon-chain/pull/314/commits/4590b697abd71fce26f17d39eeffe6a64994e4d0 is, yes, to your point, substantially operating by reducing memory churn. It's not especially accidentally-quadratic, just, slow, because of the memory issues.

So it's nuanced, and what you point out is part of it, and would indeed be interesting to check, but isn't the whole explanation.